% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_vectors.R
\name{$.circular}
\alias{$.circular}
\title{extract attribute}
\usage{
\method{$}{circular}(x, attr)
}
\arguments{
\item{x}{a circular vector}

\item{attr}{attribute to extract}
}
\description{
extract attribute
}
