% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{$.segment}
\alias{$.segment}
\title{extract
Method for accessing attributes of a 'segment' object with the $ operator}
\usage{
\method{$}{segment}(x, attr)
}
\arguments{
\item{x}{x}

\item{attr}{attribute}
}
\value{
logical
}
\description{
extract
Method for accessing attributes of a 'segment' object with the $ operator
}
