% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_vectors.R
\name{circular}
\alias{circular}
\title{circular
Constructor function for 'circular' S3 class}
\usage{
circular(x)
}
\arguments{
\item{x}{a one dimentional vector}
}
\value{
a vector of class circular
}
\description{
circular
Constructor function for 'circular' S3 class
}
\examples{
circular(1:3)
}
