% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{intersect.segment}
\alias{intersect.segment}
\title{intersect segments
find intersection between 2 segments}
\usage{
intersect.segment(s1, s2)
}
\arguments{
\item{s1}{segment 1}

\item{s2}{segment 2}
}
\value{
x,y vector
}
\description{
intersect segments
find intersection between 2 segments
}
