% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_utils.R
\name{mat_expand}
\alias{mat_expand}
\title{mat_expand
expand matrix in either directions}
\usage{
mat_expand(x, directions = c("L", "R", "U", "D"), n, fill)
}
\arguments{
\item{x}{matrix}

\item{directions}{one or more directions of "L","R","U","D"}

\item{n}{number of extra rows/columns to add}

\item{fill}{fill extra rows with}
}
\value{
matrix
}
\description{
mat_expand
expand matrix in either directions
}
\examples{
m <- matrix(1:4, nrow=2)
mat_expand(m, directions = c("U","D"), n=2, fill=0)
}
