% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_utils.R
\name{mat_mirror}
\alias{mat_mirror}
\title{mat_mirror
mirror matrix vertically, horizontally, or both}
\usage{
mat_mirror(x, axis = c("V", "H", "VH", "HV"))
}
\arguments{
\item{x}{matrix}

\item{axis}{one of "V", "H", "VH","HV". V= vertical, H = Horizontal}
}
\value{
matrix
}
\description{
mat_mirror
mirror matrix vertically, horizontally, or both
}
\examples{
m <- matrix(1:4, nrow=2)
mat_mirror(m, "H")
}
