% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_utils.R
\name{mat_rotate}
\alias{mat_rotate}
\title{mat_rotate
rotate matrix 90, 180, 270 ... clockwise}
\usage{
mat_rotate(x, n)
}
\arguments{
\item{x}{matrix}

\item{n}{number of clockwise 90° rotations}
}
\value{
matrix
}
\description{
mat_rotate
rotate matrix 90, 180, 270 ... clockwise
}
\examples{
m <- matrix(1:4, nrow=2)
mat_rotate(m, n=3)
}
