% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_aoc_input_with}
\alias{read_aoc_input_with}
\title{read_aoc_input_with
read AoC input with a given function}
\usage{
read_aoc_input_with(year, day, .fun, ...)
}
\arguments{
\item{year}{year}

\item{day}{day}

\item{.fun}{function to read input with}

\item{...}{extra args to pass to .fun}
}
\value{
depends on .fun
}
\description{
read_aoc_input_with
read AoC input with a given function
}
