% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{seg_include}
\alias{seg_include}
\title{Title
check if a position is found on a segment}
\usage{
seg_include(segment, x, y, esclude_extremes = FALSE)
}
\arguments{
\item{segment}{segment}

\item{x}{x}

\item{y}{y}

\item{esclude_extremes}{logical}
}
\value{
logical
}
\description{
Title
check if a position is found on a segment
}
