% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{seg_include.segment}
\alias{seg_include.segment}
\title{seg_include
check if a position is found on a segment
must fix. doesn't work for vertical and horizontal segments}
\usage{
\method{seg_include}{segment}(segment, x, y, esclude_extremes = FALSE)
}
\arguments{
\item{segment}{segment}

\item{x}{x}

\item{y}{y}

\item{esclude_extremes}{logical}
}
\value{
logical
}
\description{
seg_include
check if a position is found on a segment
must fix. doesn't work for vertical and horizontal segments
}
