% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment}
\alias{segment}
\title{segment
Constructor function for 'segments' S3 class}
\usage{
segment(x)
}
\arguments{
\item{x}{a numeric vector of length 4, with start/end coordinates of the segment: x1, y1, x2, y2}
}
\value{
segment
}
\description{
segment
Constructor function for 'segments' S3 class
}
\examples{
segment(1:4)
}
