% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroindex_vectors.R
\name{[.zeroindex}
\alias{[.zeroindex}
\title{subset}
\usage{
\method{[}{zeroindex}(x, i)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{index}
}
\description{
subset
}
