% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_vectors.R
\name{[<-.circular}
\alias{[<-.circular}
\title{assignment by index}
\usage{
\method{[}{circular}(x, i) <- value
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{index}

\item{value}{assign to value}
}
\description{
assignment by index
}
