% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroindex_vectors.R
\name{zeroindex}
\alias{zeroindex}
\title{zeroindex
Constructor function for 'zeroindex' S3 class}
\usage{
zeroindex(x)
}
\arguments{
\item{x}{a one dimentional vector}
}
\value{
a vector of class zeroindex
}
\description{
zeroindex
Constructor function for 'zeroindex' S3 class
}
\examples{
zeroindex(1:3)
}
